Cleanup
- Split tasks into separate folders/files
- Remove un-commented code
- Factor out application setup
- Add static client as part of application setup
